% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{getCalendar}
\alias{getCalendar}
\title{Get new calendar objectb}
\usage{
getCalendar(calstr)
}
\arguments{
\item{calstr}{Character variable identifying desired calendar}
}
\value{
A external pointer classed as S3 class 'qlcalendar'
}
\description{
Get new calendar
}
\details{
This function returns an external pointer, classed as small S3 helper class,
to a new QuantLib Calendar object identified by the calendar string.
}
\examples{
xp <- getCalendar("UnitedStates/NYSE")
xp  # invokes the print method
xp2 <- getCalendar("Canada/TSX")
xp2 # invokes the print method
}
\seealso{
\code{setCalendar}
}
