% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{isHoliday}
\alias{isHoliday}
\title{Test for holidays}
\usage{
isHoliday(dates = NULL, xp = NULL)
}
\arguments{
\item{dates}{An optional Date vector with dates to be examined, if missing the
current day is used}

\item{xp}{An optional calendar object, if missing the default instance is used}
}
\value{
A logical vector indicating which dates are holidays
}
\description{
Test a vector of dates for holiday
}
\details{
This function takes a vector of dates and returns a logical vector
of the same length indicating at each position whether the corresponding
date is a holiday in the currently active (global) calendar.
}
\examples{
isHoliday(Sys.Date()+0:6)
}
