% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{isWeekend}
\alias{isWeekend}
\title{Test for weekends}
\usage{
isWeekend(dates = NULL, xp = NULL)
}
\arguments{
\item{dates}{An optional Date vector with dates to be examined, if missing the
current day is used}

\item{xp}{An optional calendar object, if missing the default instance is used}
}
\value{
A logical vector indicating which dates are weekends
}
\description{
Test a vector of dates for weekends
}
\details{
This function takes a vector of dates and returns a logical vector
of the same length indicating at each position whether the corresponding
date is a weekend in the currently active (global) calendar.
}
\examples{
isWeekend(Sys.Date()+0:6)
}
