\name{qlcal-package}
\alias{qlcal-package}
\alias{qlcal}
\docType{package}
\title{\packageTitle{qlcal}}
\description{\packageDescription{qlcal}}
\details{The DESCRIPTION file: \packageDESCRIPTION{qlcal}}
\section{Package Content}{\packageIndices{qlcal}}
\author{\packageAuthor{qlcal}}
\section{Maintainer}{\packageMaintainer{qlcal}}
\references{\url{https://www.quantlib.org/}}
\keyword{package}
