% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{setCalendar}
\alias{setCalendar}
\title{Set a calendar}
\usage{
setCalendar(calstr)
}
\arguments{
\item{calstr}{A character variable containing the market for which a calendar
is to be set}
}
\value{
Nothing is returned but the global state is changed
}
\description{
Set a calendar
}
\details{
This function sets the default calendar to the given market or country convention.
Note that additional calendar objects can be created with the \code{getCalendar}
function.
}
\examples{
setCalendar("UnitedStates/NYSE")
setCalendar("Canada/TSX")
}
\seealso{
\code{getCalendar}
}
